refactor(common/device-core): make managerApi domain directory
