+++ libs/ledger-live-common/src/device-core/managerApi/entities/DeviceInfoEntity.ts
