+++ libs/ledger-live-common/src/device-core/managerApi/repositories/ManagerApiRepository.ts
