+++ libs/ledger-live-common/src/device-core/managerApi/use-cases/getLatestFirmwareForDevice.ts
